% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LoanPayments.R
\name{LoanPayments}
\alias{LoanPayments}
\title{Function to calculate loan repayments
based on:
a) loan amount (present value)
b) annual interest rate
c) number of monthly payment periods
Function returns monthly payment
Useful for e.g. working out repayment mortgage payments}
\usage{
LoanPayments(pv, r, n)
}
\arguments{
\item{pv}{a present value}

\item{r}{a annual compound interest rate}

\item{n}{a number of payments e.g.  20 years = 240 payments}
}
\value{
p the amount of each monthly payment
}
\description{
Function to calculate loan repayments
based on:
a) loan amount (present value)
b) annual interest rate
c) number of monthly payment periods
Function returns monthly payment
Useful for e.g. working out repayment mortgage payments
}
\details{
This function uses a formula to calculate the monthly repayments
of a loan over a number of months
}
\author{
Michael Fortune
}

